Add tasklet and support for taking possession of an existing simulator
Takes possession of any existing region simulator bin folder and all regions controlled by said simulator. This is doe by the simple expedient of copying the whole bin folder, sending CTRL-C to the old process if it's online and then waitign until the porcess goes away, then starting up the new copy.

Doesn't require the simulator to be offline first, but it's generally a good idea in case anything that should be in the region .xml files has changed.
It DOES require that the region not auto resstart when shut down!